[FIX] 메인 보호자 화면 새로고침 시 도장판 규격 작아지는 버그 수정
