+++ app/src/main/res/layout/fragment_progress.xml
@@ -9,73 +9,76 @@
             type="com.polzzak_android.presentation.main.protector.StampViewModel" />
     </data>
 
-    <androidx.core.widget.NestedScrollView
+    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+        android:id="@+id/stampListRefresh"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:fillViewport="true">
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/selectContainer">
 
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/mainContainer"
+
+        <androidx.core.widget.NestedScrollView
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@color/gray_100">
+            android:layout_height="match_parent"
+            android:fillViewport="true"
+            android:scrollbars="none">
 
-            <!-- 유저 필터링 -->
             <androidx.constraintlayout.widget.ConstraintLayout
-                android:id="@+id/selectContainer"
-                android:layout_width="wrap_content"
+                android:id="@+id/mainContainer"
+                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:paddingHorizontal="15dp"
-                android:paddingVertical="20dp"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintHorizontal_bias="0"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent">
+                android:background="@color/gray_100">
 
-                <TextView
-                    android:id="@+id/selectTxt"
-                    style="@style/title.22.700"
+                <!-- 유저 필터링 -->
+                <androidx.constraintlayout.widget.ConstraintLayout
+                    android:id="@+id/selectContainer"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:textColor="@color/gray_800"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintEnd_toStartOf="@id/selectBtn"
-                    app:layout_constraintHorizontal_chainStyle="spread_inside"
+                    android:paddingHorizontal="15dp"
+                    android:paddingVertical="20dp"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintHorizontal_bias="0"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    tools:text="전체" />
+                    app:layout_constraintTop_toTopOf="parent">
 
-                <View
-                    android:id="@+id/selectBtn"
-                    android:layout_width="20dp"
-                    android:layout_height="20dp"
-                    android:layout_marginStart="11dp"
-                    android:background="@color/primary"
-                    android:padding="15dp"
-                    app:layout_constraintBottom_toBottomOf="@id/selectTxt"
-                    app:layout_constraintStart_toEndOf="@id/selectTxt"
-                    app:layout_constraintTop_toTopOf="@id/selectTxt" />
+                    <TextView
+                        android:id="@+id/selectTxt"
+                        style="@style/title.22.700"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:textColor="@color/gray_800"
+                        app:layout_constraintBottom_toBottomOf="parent"
+                        app:layout_constraintEnd_toStartOf="@id/selectBtn"
+                        app:layout_constraintHorizontal_chainStyle="spread_inside"
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"
+                        tools:text="전체" />
 
-            </androidx.constraintlayout.widget.ConstraintLayout>
+                    <View
+                        android:id="@+id/selectBtn"
+                        android:layout_width="20dp"
+                        android:layout_height="20dp"
+                        android:layout_marginStart="11dp"
+                        android:background="@color/primary"
+                        android:padding="15dp"
+                        app:layout_constraintBottom_toBottomOf="@id/selectTxt"
+                        app:layout_constraintStart_toEndOf="@id/selectTxt"
+                        app:layout_constraintTop_toTopOf="@id/selectTxt" />
 
-            <!-- 도장판 목록 -->
-            <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-                android:id="@+id/stampListRefresh"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/selectContainer">
+                </androidx.constraintlayout.widget.ConstraintLayout>
 
+                <!-- 도장판 목록 -->
                 <androidx.recyclerview.widget.RecyclerView
                     android:id="@+id/stampListRc"
                     android:layout_width="match_parent"
                     android:layout_height="0dp"
+                    app:layout_constraintTop_toBottomOf="@id/selectContainer"
                     tools:listitem="@layout/item_stamp_yes" />
 
-            </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
+            </androidx.constraintlayout.widget.ConstraintLayout>
 
-        </androidx.constraintlayout.widget.ConstraintLayout>
+        </androidx.core.widget.NestedScrollView>
 
-    </androidx.core.widget.NestedScrollView>
+    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
 </layout>
